Various SQL schema fixes
git-svn-id: http://svn.a2billing.net/svn/asterisk2billing/trunk@246 9d51a8f0-6521-42ea-932e-d822aa02ce2f

+++ DataBase/mysql-5.x/a2billing-mysql-schema-MYSQL.5.X-v1.3.0.sql
@@ -39,7 +39,7 @@ CREATE TABLE cc_did (
     id_cc_country 					INT NOT NULL,    
     activated 						INT DEFAULT '1' NOT NULL,
     reserved 						INT DEFAULT '0',
-    iduser 							INT DEFAULT '0' NOT NULL,
+    iduser 						BIGINT DEFAULT '0' NOT NULL,
     did 							CHAR(50) NOT NULL,
     creationdate  					TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     startingdate  					TIMESTAMP,
